requirements.txt
pandas
openpyxl
openai
pinecone-client
replicate

black
flake8
isort
